# ApiCenter/ApiDoc.yaml
swagger: '2.0'
info:
  version: '1.0-SNAPSHOT'
  title: YiWu-Server
  description: yw数据服务
host: 127.0.0.1:8081
#basePath: /YiWu-Server
schemes:
  - http
paths:
  /idle:
    post:
      tags:
        - idle
      description: 上传商品信息
      parameters:
        - in: body
          name: IdleModel
          description: 上传商品入参模型
          schema:
            type: object
            $ref: "#/definitions/IdleModel"
      responses:
        200:
          description: 请求成功
        500:
            $ref: "#/responses/Standard500ErrorResponse"
        404:
            $ref: "#/responses/Standard404ErrorResponse"
    put:
      tags:
        - idle
      description: 上传商品信息
      parameters:
        - in: body
          name: IdleModel
          description: 上传商品入参模型
          schema:
            type: object
            $ref: "#/definitions/IdleModel"
      responses:
        200:
          description: 请求成功
        500:
            $ref: "#/responses/Standard500ErrorResponse"
        404:
            $ref: "#/responses/Standard404ErrorResponse"
securityDefinitions:
  access_token:
    type: apiKey
    name: Authorization
    in: header
definitions:
    Error:
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    IdleModel:
          required:
            - name
            - typeId
            - price
            - num
            - extent
          properties:
            name:
              type: string
              description: 商品名称
            description:
              type: string
              description: 商品描述
            price:
              type: number
              format: double
              description: 商品单价
            num:
              type: integer
              description: 商品数量
            typeId:
              type: string
              description: 商品类型id
            productionDate:
              type: string
              format: date
              description: 商品生产日期
            productionPlace:
              type: string
              description: 商品产地
            guaranteeDate:
              type: string
              format: date
              description: 商品保质日期
            extent:
              type: integer
              description: 商品新旧程度（0-100）
            shareValue:
              type: integer
              description: 商品共享值
            shareCycle:
              type: integer
              description: 商品共享周期
            thumbnail:
              type: string
              description: 商品缩略图
responses:
  Standard500ErrorResponse:
    description: 请求失败
    schema:
      $ref: "#/definitions/Error"
  Standard404ErrorResponse:
    description: 页面不存在
parameters:
  pageSize:
    name: pageSize
    in: query
    description: 每页大小
    type: integer
    format: int32
    required: false
  page:
    in: query
    name: page
    description: 页码
    type: integer
    format: int32
    required: false